requirements.txt
customtkinter==5.2.1
darkdetect==0.8.0
packaging==23.2
Pillow==10.1.0
pip==23.2.1
PyMySQL==1.1.0
setuptools==68.2.2
wheel==0.41.2